+++ electron/src/electron-utils/utilsFile.ts
@@ -2,6 +2,7 @@ export class UtilsFile {
   pathDB = 'Databases';
   Path: any = null;
   NodeFs: any = null;
+  NodeFetch: any = null;
   JSZip: any = null;
   Os: any = null;
   Electron: any = null;
@@ -15,6 +16,7 @@ export class UtilsFile {
   constructor() {
     this.Path = require('path');
     this.NodeFs = require('fs');
+    this.NodeFetch = require('node-fetch');
     this.Os = require('os');
     this.JSZip = require('jszip');
     this.Electron = require('electron');
@@ -63,6 +65,17 @@ export class UtilsFile {
         console.log('other operating system');
     }
     console.log(`&&& Databases path: ${this.pathDB}`);
+  }
+  /**
+   * GetExtName
+   * @param filePath
+   * @returns
+   */
+  public getExtName(filePath: string): string {
+    return this.Path.extname(filePath);
+  }
+  public getBaseName(filePath: string): string {
+    return this.Path.basename(filePath, this.Path.extname(filePath));
   }
   /**
    * IsPathExists
@@ -122,6 +135,18 @@ export class UtilsFile {
     return retPath;
   }
   /**
+   * GetCachePath
+   * get the database cache folder path
+   */
+  public getCachePath(): string {
+    let retPath = '';
+    const databasePath = this.getDatabasesPath();
+    retPath = this.Path.join(databasePath, 'cache');
+    const retB: boolean = this._createFolderIfNotExists(retPath);
+    if (!retB) retPath = '';
+    return retPath;
+  }
+  /**
    * GetAssetsDatabasesPath
    * get the assets databases folder path
    */
@@ -143,10 +168,12 @@ export class UtilsFile {
   public setPathSuffix(db: string): string {
     let toDb: string = db;
     const ext = '.db';
-    const sep = this.Path.sep;
-    if (db.substring(db.length - 3) === ext) {
-      if (!db.includes('SQLite.db')) {
-        toDb = db.slice(db.lastIndexOf(sep) + 1, -3) + 'SQLite.db';
+    const dirName = this.Path.dirname(db);
+    const baseName = this.getBaseName(db);
+    if (this.getExtName(db) === ext) {
+      if (!baseName.includes('SQLite')) {
+        const dbName = `${baseName}SQLite`;
+        toDb = `${this.Path.join(dirName, dbName)}${ext}`;
       }
     }
     return toDb;
@@ -160,7 +187,7 @@ export class UtilsFile {
     const filenames = this.NodeFs.readdirSync(path);
     const dbs: string[] = [];
     filenames.forEach((file: string) => {
-      if (this.Path.extname(file) == '.db' || this.Path.extname(file) == '.zip')
+      if (this.getExtName(file) == '.db' || this.getExtName(file) == '.zip')
         dbs.push(file);
     });
     return Promise.resolve(dbs);
@@ -185,8 +212,12 @@ export class UtilsFile {
    * @param db
    * @param overwrite
    */
-  public async unzipDatabase(db: string, overwrite: boolean): Promise<void> {
-    const pZip: string = this.Path.join(this.getAssetsDatabasesPath(), db);
+  public async unzipDatabase(
+    db: string,
+    fPath: string,
+    overwrite: boolean,
+  ): Promise<void> {
+    const pZip: string = this.Path.join(fPath, db);
     // Read the Zip file
     this.NodeFs.readFile(pZip, (err: any, data: any) => {
       if (err) {
@@ -373,10 +404,36 @@ export class UtilsFile {
           return Promise.reject('RenameFilePath: ' + `${err}`);
         }
       } else {
-        return Promise.reject('RenameFilePath: filePath ' + 'does not exist');
+        return Promise.reject(`RenameFilePath: ${filePath} does not exist`);
       }
     } else {
       return Promise.reject('RenameFilePath: filePath not found');
+    }
+  }
+  public async moveDatabaseFromCache(): Promise<void> {
+    const cachePath: string = this.getCachePath();
+    const databasePath: string = this.getDatabasesPath();
+    const dbCacheList: string[] = await this.getFileList(cachePath);
+    for (const name of dbCacheList) {
+      const ext: string = this.getExtName(name);
+      const fromDBName: string = this.Path.join(cachePath, name);
+      if (ext === '.db') {
+        const pDb: string = this.setPathSuffix(
+          this.Path.join(databasePath, name),
+        );
+        try {
+          await this.renameFilePath(fromDBName, pDb);
+        } catch (err) {
+          return Promise.reject('moveDatabaseFromCache: ' + `${err}`);
+        }
+      }
+      if (ext === '.zip') {
+        try {
+          await this.deleteFilePath(fromDBName);
+        } catch (err) {
+          return Promise.reject('moveDatabaseFromCache: ' + `${err}`);
+        }
+      }
     }
   }
   /**
@@ -411,6 +468,26 @@ export class UtilsFile {
         `RestoreFileName: ${mFileName} ` + 'does not exist',
       );
     }
+  }
+  /**
+   * DownloadFileFromHTTP
+   * @param url
+   * @param path
+   */
+  public async downloadFileFromHTTP(
+    url: string,
+    pathFolder: string,
+  ): Promise<void> {
+    const res: any = await this.NodeFetch(url);
+    const ext: string = this.getExtName(url);
+    const dbName: string = this.getBaseName(url);
+    const filePath = `${this.Path.join(pathFolder, dbName)}${ext}`;
+    const fileStream: any = this.NodeFs.createWriteStream(filePath);
+    await new Promise((resolve, reject) => {
+      res.body.pipe(fileStream);
+      res.body.on('error', reject);
+      fileStream.on('finish', resolve);
+    });
   }
   /**
    * CreateFolderIfNotExists