+++ ios/Plugin/CapacitorSQLitePlugin.m
@@ -44,4 +44,5 @@ CAP_PLUGIN(CapacitorSQLitePlugin, "CapacitorSQLite",
            CAP_PLUGIN_METHOD(setEncryptionSecret, CAPPluginReturnPromise);
            CAP_PLUGIN_METHOD(changeEncryptionSecret, CAPPluginReturnPromise);
            CAP_PLUGIN_METHOD(clearEncryptionSecret, CAPPluginReturnPromise);
+           CAP_PLUGIN_METHOD(getFromHTTPRequest, CAPPluginReturnPromise);
 )